Tue, Mar 19, 2024, 10:41 PM +06:00

+++ module 14.5/project_14_5/app1/migrations/0001_initial.py
@@ -15,7 +15,7 @@ class Migration(migrations.Migration):
             fields=[
                 ("ID", models.IntegerField(primary_key=True, serialize=False)),
                 ("name", models.CharField(max_length=10)),
-                ("balance", models.BigIntegerField()),
+                ("balance", models.IntegerField()),
                 ("login", models.BooleanField(default=True)),
                 ("about", models.TextField(default=None, max_length=50)),
                 ("date", models.DateField()),
