+++ module 14.5/project_14_5/app1/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class MyDBFormClass(models.Model):
     ID = models.IntegerField(primary_key=True)
     name = models.CharField(max_length=10)
-    balance = models.BigIntegerField()
+    balance = models.IntegerField()
     login = models.BooleanField(default=True)
     about = models.TextField(max_length=50, default=None)
     date = models.DateField()
